# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Traffic safety community portal backend.",
        "title": "Obvio Community Portal API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/portals": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "portals"
                ],
                "summary": "List all portals",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/portal.Portal"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "portals"
                ],
                "summary": "Create a portal",
                "parameters": [
                    {
                        "description": "Portal data",
                        "name": "portal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/portal.Portal"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/portal.Portal"
                        }
                    }
                }
            }
        },
        "/portals/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "portals"
                ],
                "summary": "Get portal by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Portal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/portal.Portal"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "portals"
                ],
                "summary": "Update a portal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Portal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Portal data",
                        "name": "portal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/portal.Portal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/portal.Portal"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "portals"
                ],
                "summary": "Delete a portal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Portal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/portals/{id}/trials": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trials"
                ],
                "summary": "List trials for a portal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Portal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/trial.Trial"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trials"
                ],
                "summary": "Create a trial for a portal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Portal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Trial data",
                        "name": "trial",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/trial.Trial"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/trial.Trial"
                        }
                    }
                }
            }
        },
        "/trials/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trials"
                ],
                "summary": "Get trial by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Trial ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/trial.Trial"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trials"
                ],
                "summary": "Update a trial",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Trial ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Trial data",
                        "name": "trial",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/trial.Trial"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/trial.Trial"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "trials"
                ],
                "summary": "Delete a trial",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Trial ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "portal.Portal": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "trial.Trial": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "portal_id": {
                    "type": "integer"
                },
                "start_date": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}